requirements.txt
pyaudio
pynput
webrtcvad-wheels
python-dotenv
requests
pyinstaller